+++ Assets/Script/Camera.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: c04bc21b50be8f942b762a4b596b9341
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
