+++ Assets/Script/CameraController.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 33f9ef1caa173c743b6fd2bd68cc0c0e
+guid: b3023e18ade0cd446b7777fd85d95017
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
