--- Assets/Script/ZoomCamera.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 4fe4df881465bb047976b8a8ef2babd9
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
