+++ ProjectSettings/ProjectVersion.txt
@@ -1,2 +1,2 @@
-m_EditorVersion: 2021.3.5f1
-m_EditorVersionWithRevision: 2021.3.5f1 (40eb3a945986)
+m_EditorVersion: 2021.3.4f1
+m_EditorVersionWithRevision: 2021.3.4f1 (cb45f9cae8b7)
